requirements.txt
Django==5.2.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
google-generativeai
python-dotenv
python-dateutil
psycopg2-binary
fastapi==0.104.1
uvicorn[standard]==0.24.0
selenium==4.15.2
beautifulsoup4==4.12.2
lxml==4.9.3
apscheduler==3.10.4
pydantic==2.5.0
yfinance==0.2.66
transformers==4.57.3
torch==2.9.1
reportlab==4.4.5